# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.11.0",
    "title": "SiSMoChat API",
    "description": "API for SiSMoChat - A secure chat application focused on child safety with parental control."
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT token. Format: \"Bearer {token}\""
    }
  },
  "paths": {
    "/auth/parent": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Authenticate parent",
        "description": "Login with email and password, returns a JWT token.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Parent"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT token returned"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/parents/": {
      "post": {
        "tags": [
          "Messages"
        ],
        "summary": "Send a message",
        "description": "Send an encrypted message to another user.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Message"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Message sent, returns messageID"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/parents/password": {
      "patch": {
        "tags": [
          "Parents"
        ],
        "summary": "Change parent password",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "oldPassword": {
                  "example": "any"
                },
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Old password incorrect"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/parents/me/connections/sent": {
      "get": {
        "tags": [
          "Parents"
        ],
        "summary": "Get sent connection requests",
        "description": "",
        "responses": {
          "200": {
            "description": "List of sent requests with status"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Parent not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/parents/me/connections/pending": {
      "get": {
        "tags": [
          "Parents"
        ],
        "summary": "Get pending connection requests for approval",
        "description": "",
        "responses": {
          "200": {
            "description": "List of pending connection requests"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Parent not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/parents/{email}/children": {
      "get": {
        "tags": [
          "Parents"
        ],
        "summary": "Get children of a parent by email",
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of children (nick only)"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Parent not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/parents/{email}": {
      "get": {
        "tags": [
          "Parents"
        ],
        "summary": "Get parent by email",
        "description": "",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Parent found"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Parent not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/parents/reset-request": {
      "post": {
        "tags": [
          "Parents"
        ],
        "summary": "Request password reset OTP",
        "description": "Sends a 6-digit OTP to the parent email. Does not reveal if email exists.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset email sent (if account exists)"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/parents/reset": {
      "post": {
        "tags": [
          "Parents"
        ],
        "summary": "Reset password with OTP",
        "description": "Validates the 6-digit OTP and sets a new password. Max 5 attempts.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "otp": {
                  "example": "any"
                },
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Password reset successful"
          },
          "400": {
            "description": "Invalid or expired OTP"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Parent": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "parent@example.com"
        },
        "pwd": {
          "type": "string",
          "example": "securepassword"
        }
      }
    },
...
    "Message": {
      "type": "object",
      "properties": {
        "to": {
          "type": "string",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "message": {
          "type": "string",
          "example": "encrypted message content"
        }
      }
    },
...
  }
}